tests/wpt/web-platform-tests/dom/events/Event-dispatch-click.tentative.html

<!doctype html>
<title>Clicks on input element</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=dump style=display:none></div>
<script>
var dump = document.getElementById("dump")

test(t => {
  const input = document.createElement("input");
  input.type = "checkbox";
  input.disabled = true;
  const label = document.createElement("label");
  label.append(input);
  dump.append(label);
  label.click();
  assert_false(input.checked);
}, "disabled checkbox should not be checked from label click");

test(t => {
  const input = document.createElement("input");
  input.type = "radio";
  input.disabled = true;
  const label = document.createElement("label");
  label.append(input);
  dump.append(label);
  label.click();
  assert_false(input.checked);
}, "disabled radio should not be checked from label click");

test(t => {
  const input = document.createElement("input");
  input.type = "checkbox";
  input.disabled = true;
  const label = document.createElement("label");
  label.append(input);
  dump.append(label);
  label.dispatchEvent(new MouseEvent("click"));
  assert_false(input.checked);
}, "disabled checkbox should not be checked from label click by dispatchEvent");

test(t => {
  const input = document.createElement("input");
  input.type = "radio";
  input.disabled = true;
  const label = document.createElement("label");
  label.append(input);
  dump.append(label);
  label.dispatchEvent(new MouseEvent("click"));
  assert_false(input.checked);
}, "disabled radio should not be checked from label click by dispatchEvent");
</script>
